# [hooks_runner] No hooks performance trace